+++ pypgx/cli/prepare_depth_of_coverage.py
@@ -12,6 +12,8 @@ description = f"""
 
 Input BAM files must be specified with either '--bam' or '--fn', but it's an error to use both.
 
+By default, the input data is assumed to be WGS. If it's targeted sequencing, you must provide a BED file with '--bed' to indicate probed regions.
+
 Usage examples:
   $ fuc {fuc.api.common._script_name()} depth-of-coverage.tsv --bam A.bam B.bam
   $ fuc {fuc.api.common._script_name()} depth-of-coverage.tsv --fn bam.list
@@ -46,9 +48,14 @@ def create_parser(subparsers):
         default='GRCh37',
         help="Reference genome assembly (default: 'GRCh37') (choices: 'GRCh37', 'GRCh38')."
     )
+    parser.add_argument(
+        '--bed',
+        metavar='PATH',
+        help='BED file.'
+    )
 
 def main(args):
     cf = utils.prepare_depth_of_coverage(
-        bam=args.bam, fn=args.fn, assembly=args.assembly,
+        bam=args.bam, fn=args.fn, assembly=args.assembly, bed=args.bed
     )
     cf.to_file(args.depth_of_coverage)