+++ pypgx/cli/run_ngs_pipeline.py
@@ -10,7 +10,7 @@ description = f"""
 #########################################
 
 Usage examples:
-  $ pypgx {fuc.api.common._script_name()} CYP2D6 CYP2D6-pipeline --vcf input.vcf --tsv input.tsv --control-statistcs control-statistics-VDR.zip
+  $ pypgx {fuc.api.common._script_name()} CYP2D6 CYP2D6-pipeline --variants variants.vcf --depth-of-coverage depth-of-coverage.tsv --control-statistcs control-statistics-VDR.zip
 """
 
 def create_parser(subparsers):
@@ -29,14 +29,14 @@ def create_parser(subparsers):
         help='Output directory.'
     )
     parser.add_argument(
-        '--vcf',
+        '--variants',
         metavar='PATH',
-        help='VCF file.'
+        help='VCF file (zipped or unzipped).'
     )
     parser.add_argument(
-        '--tsv',
+        '--depth-of-coverage',
         metavar='PATH',
-        help='TSV file containing read depth (zipped or unzipped).'
+        help='Depth of coverage file (zipped or unzipped).'
     )
     parser.add_argument(
         '--control-statistics',
@@ -44,14 +44,14 @@ def create_parser(subparsers):
         help='Archive file with the semandtic type SampleTable[Statistcs].'
     )
     parser.add_argument(
-        '--force',
-        action='store_true',
-        help='Overwrite output directory if it already exists.'
-    )
-    parser.add_argument(
         '--panel',
         metavar='PATH',
         help='Reference haplotype panel. By default, the 1KGP panel is used.'
+    )
+    parser.add_argument(
+        '--force',
+        action='store_true',
+        help='Overwrite output directory if it already exists.'
     )
     parser.add_argument(
         '--do-not-plot-copy-number',
@@ -66,8 +66,10 @@ def create_parser(subparsers):
 
 def main(args):
     pipeline.run_ngs_pipeline(
-        args.gene, args.output, vcf=args.vcf, panel=args.panel, tsv=args.tsv,
-        control_statistics=args.control_statistics, force=args.force,
+        args.gene, args.output, variants=args.variants,
+        depth_of_coverage=args.depth_of_coverage,
+        control_statistics=args.control_statistics,
+        panel=args.panel, force=args.force,
         do_not_plot_copy_number=args.do_not_plot_copy_number,
         do_not_plot_allele_fraction=args.do_not_plot_allele_fraction
     )